+++ lib/util/web/web.dart
@@ -29,6 +29,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart'
     show NotificationResponse, NotificationResponseType;
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+import 'package:hotkey_manager/hotkey_manager.dart';
 import 'package:mutex/mutex.dart';
 import 'package:platform_detect/platform_detect.dart';
 import 'package:uuid/uuid.dart';
@@ -132,6 +133,10 @@ class WebUtils {
 
   /// [Mutex]es guarding the [protect] method.
   static final Map<String, Mutex> _guards = {};
+
+  /// Handlers for [HotKey]s intended to be manipulated via [bindKey] and
+  /// [unbindKey] to invoke [_handleBindKeys].
+  static final Map<HotKey, List<bool Function()>> _keyHandlers = {};
 
   /// Indicates whether device's OS is macOS or iOS.
   static bool get isMacOS =>
@@ -827,6 +832,73 @@ class WebUtils {
     final info = await DeviceInfoPlugin().webBrowserInfo;
     return info.userAgent ??
         '${Config.userAgentProduct}/${Config.userAgentVersion}';
+  }
+
+  /// Binds the [handler] to be invoked on the [key] presses.
+  static Future<void> bindKey(HotKey key, bool Function() handler) async {
+    if (_keyHandlers.isEmpty) {
+      HardwareKeyboard.instance.addHandler(_handleBindKeys);
+    }
+
+    final List<bool Function()>? contained = _keyHandlers[key];
+    if (contained == null) {
+      _keyHandlers[key] = [handler];
+    } else {
+      contained.add(handler);
+    }
+  }
+
+  /// Unbinds the [handler] from the [key].
+  static Future<void> unbindKey(HotKey key, bool Function() handler) async {
+    final list = _keyHandlers[key];
+    list?.remove(handler);
+
+    if (list?.isEmpty == true) {
+      _keyHandlers.remove(key);
+    }
+
+    if (_keyHandlers.isEmpty) {
+      HardwareKeyboard.instance.removeHandler(_handleBindKeys);
+    }
+  }
+
+  /// Handles the [key] event to invoke [_keyHandlers] related to it.
+  static bool _handleBindKeys(KeyEvent key) {
+    if (key is KeyUpEvent) {
+      for (var e in _keyHandlers.entries) {
+        if (e.key.key == key.physicalKey) {
+          bool modifiers = true;
+
+          for (var m in e.key.modifiers ?? <HotKeyModifier>[]) {
+            modifiers =
+                modifiers &&
+                switch (m) {
+                  HotKeyModifier.alt => HardwareKeyboard.instance.isAltPressed,
+                  HotKeyModifier.capsLock => HardwareKeyboard.instance
+                      .isPhysicalKeyPressed(PhysicalKeyboardKey.capsLock),
+                  HotKeyModifier.control =>
+                    HardwareKeyboard.instance.isControlPressed,
+                  HotKeyModifier.fn => HardwareKeyboard.instance
+                      .isPhysicalKeyPressed(PhysicalKeyboardKey.fn),
+                  HotKeyModifier.meta =>
+                    HardwareKeyboard.instance.isMetaPressed,
+                  HotKeyModifier.shift =>
+                    HardwareKeyboard.instance.isShiftPressed,
+                };
+          }
+
+          if (modifiers) {
+            for (var f in e.value) {
+              if (f()) {
+                return true;
+              }
+            }
+          }
+        }
+      }
+    }
+
+    return false;
   }
 }